pyproject.toml
[tool.poetry]
name = "ds_envs"
version = "2020.10.23"
description = "Data science environments"
authors = ["Mads Ehrhorn <meh@venzonxt.com>"]

[tool.poetry.dependencies]
python = "3.8.5"
data-science-types = "0.2.19"
numpy = "1.18.5"
pandas = "1.1.3"
matplotlib = "3.3.2"
tensorflow = "2.3.1"
tensorflow-probability = "0.11.1"
azure-common = "1.1.25"
azure-mgmt-resource = "10.2.0"
azure-storage-blob = "12.5.0"
azure-keyvault = "4.1.0"
azureml-sdk = "1.16.0"
azure-cli-core = "2.13.0"
azureml-tensorboard = "1.16.0"
mlflow = "1.11.0"
azureml-mlflow = "1.16.0"
debugpy = "1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=1.1.0"]
build-backend = "poetry.masonry.api"
